+++ plugins/postcss-stepped-value-functions/src/index.ts
@@ -34,6 +34,7 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 			const modifiedValue = calc(decl.value, {
 				precision: 5,
 				toCanonicalUnits: true,
+				calcWrapper: true,
 			});
 			if (modifiedValue === decl.value) {
 				return;